+++ packages/use-intl/src/core/TranslationValues.tsx
@@ -1,7 +1,7 @@
 import {ReactNode} from 'react';
 
 // From IntlMessageFormat#format
-export type TranslationValue =
+export type PlainTranslationValue =
   | string
   | number
   | boolean
@@ -9,19 +9,19 @@ export type TranslationValue =
   | null
   | undefined;
 
-type TranslationValues = Record<string, TranslationValue>;
+type TranslationValues = Record<string, PlainTranslationValue>;
 
 // We could consider renaming this to `ReactRichTranslationValues` and defining
 // it in the `react` namespace if the core becomes useful to other frameworks.
 // It would be a breaking change though, so let's wait for now.
 export type RichTranslationValues = Record<
   string,
-  TranslationValue | ((chunks: ReactNode) => ReactNode)
+  PlainTranslationValue | ((chunks: ReactNode) => ReactNode)
 >;
 
 export type MarkupTranslationValues = Record<
   string,
-  TranslationValue | ((chunks: string) => string)
+  PlainTranslationValue | ((chunks: string) => string)
 >;
 
 export default TranslationValues;
