+++ packages/use-intl/src/core/TypesafeMessages.tsx
@@ -1,59 +1,18 @@
 import {ReactNode} from 'react';
-
-/**
- * Namespaces & keys
- */
-
-export type NestedKeyOf<ObjectType> = ObjectType extends object
-  ? {
-      [Property in keyof ObjectType]:
-        | `${Property & string}`
-        | `${Property & string}.${NestedKeyOf<ObjectType[Property]>}`;
-    }[keyof ObjectType]
-  : never;
-
-export type NestedValueOf<
-  ObjectType,
-  Path extends string
-> = Path extends `${infer Cur}.${infer Rest}`
-  ? Cur extends keyof ObjectType
-    ? NestedValueOf<ObjectType[Cur], Rest>
-    : never
-  : Path extends keyof ObjectType
-    ? ObjectType[Path]
-    : never;
-
-export type NamespaceKeys<ObjectType, AllKeys extends string> = {
-  [PropertyPath in AllKeys]: NestedValueOf<
-    ObjectType,
-    PropertyPath
-  > extends string
-    ? never
-    : PropertyPath;
-}[AllKeys];
-
-export type MessageKeys<ObjectType, AllKeys extends string> = {
-  [PropertyPath in AllKeys]: NestedValueOf<
-    ObjectType,
-    PropertyPath
-  > extends string
-    ? PropertyPath
-    : never;
-}[AllKeys];
+import {PlainTranslationValue} from './TranslationValues.tsx';
 
 /**
  * Params
  */
 
-type PlainParam = string | number;
 type RichTextFunction = (chunks: ReactNode) => ReactNode;
 type MarkupFunction = (chunks: string) => string;
 
 type ExtractParams<MessageString extends string> =
   MessageString extends `${string}{${infer ParamName}}${infer RestOfMessage}`
-    ? ParamName extends `${infer Name}, ${infer FormatType}`
-      ? Record<Name, InferParamType<FormatType>> & ExtractParams<RestOfMessage>
-      : Record<ParamName, PlainParam> & ExtractParams<RestOfMessage>
+    ? ParamName extends `${infer Name}, ${string}`
+      ? Record<Name, InferParamType<`${string}`>> & ExtractParams<RestOfMessage>
+      : Record<ParamName, PlainTranslationValue> & ExtractParams<RestOfMessage>
     : {};
 
 type ExtractTags<MessageString extends string> =
@@ -69,7 +28,7 @@ type InferParamType<FormatType extends string> =
       ? number
       : FormatType extends `select, ${string}`
         ? string
-        : PlainParam;
+        : PlainTranslationValue;
 
 export type MessageParams<MessageString extends string> =
   ExtractParams<MessageString> & ExtractTags<MessageString>;
