+++ packages/use-intl/src/core/createBaseTranslator.tsx
@@ -9,7 +9,7 @@ import TranslationValues, {
   MarkupTranslationValues,
   RichTranslationValues
 } from './TranslationValues.tsx';
-import {MessageKeys, NestedKeyOf, NestedValueOf} from './TypesafeMessages.tsx';
+import {MessageKeys, NestedKeyOf, NestedValueOf} from './TypesafeKeys.tsx';
 import convertFormatsToIntlMessageFormat from './convertFormatsToIntlMessageFormat.tsx';
 import {defaultGetMessageFallback, defaultOnError} from './defaults.tsx';
 import {
