+++ packages/use-intl/src/react/useTranslations.tsx
@@ -1,5 +1,5 @@
 import {Messages} from '../core/AppConfig.tsx';
-import {NamespaceKeys, NestedKeyOf} from '../core/TypesafeMessages.tsx';
+import {NamespaceKeys, NestedKeyOf} from '../core/TypesafeKeys.tsx';
 import type createTranslator from '../core/createTranslator.tsx';
 import useIntlContext from './useIntlContext.tsx';
 import useTranslationsImpl from './useTranslationsImpl.tsx';
